renamed sdk to scripting-environment
