+++ packages/insomnia-scripting-environment/src/objects/__tests__/auth.test.ts
