+++ packages/insomnia-scripting-environment/src/objects/__tests__/cookies.test.ts
