+++ packages/insomnia-scripting-environment/src/objects/__tests__/environments.test.ts
