+++ packages/insomnia-scripting-environment/src/objects/__tests__/headers.test.ts
