+++ packages/insomnia-scripting-environment/src/objects/__tests__/properties.test.ts
