+++ packages/insomnia-scripting-environment/src/objects/__tests__/proxy-configs.test.ts
