+++ packages/insomnia-scripting-environment/src/objects/__tests__/variables.test.ts
