+++ packages/insomnia-scripting-environment/src/objects/async_objects.ts
