+++ packages/insomnia-scripting-environment/src/objects/auth.ts
