+++ packages/insomnia-scripting-environment/src/objects/collection.ts
