+++ packages/insomnia-scripting-environment/src/objects/console.ts
