+++ packages/insomnia-scripting-environment/src/objects/cookies.ts
