+++ packages/insomnia-scripting-environment/src/objects/environments.ts
