+++ packages/insomnia-scripting-environment/src/objects/folders.ts
