+++ packages/insomnia-scripting-environment/src/objects/headers.ts
