+++ packages/insomnia-scripting-environment/src/objects/index.ts
