+++ packages/insomnia-scripting-environment/src/objects/insomnia.ts
