+++ packages/insomnia-scripting-environment/src/objects/interfaces.ts
