+++ packages/insomnia-scripting-environment/src/objects/interpolator.ts
