+++ packages/insomnia-scripting-environment/src/objects/properties.ts
