+++ packages/insomnia-scripting-environment/src/objects/utils.ts
