+++ packages/insomnia-scripting-environment/src/objects/variables.ts
