+++ packages/insomnia/src/models/runner-test-result.ts
@@ -1,5 +1,4 @@
-import type { RequestTestResult } from 'insomnia-sdk';
-
+import type { RequestTestResult } from '../../../insomnia-scripting-environment/src/objects';
 import { database as db } from '../common/database';
 import type { RunnerSource } from '../ui/routes/request';
 import type { BaseModel } from './index';
