+++ packages/insomnia/src/ui/components/panes/request-test-result-pane.tsx
@@ -1,8 +1,8 @@
 import crypto from 'crypto';
-import type { RequestTestResult } from 'insomnia-sdk';
 import React, { type FC, useState } from 'react';
 import { Toolbar } from 'react-aria-components';
 
+import type { RequestTestResult } from '../../../../../insomnia-scripting-environment/src/objects';
 import { fuzzyMatch } from '../../../common/misc';
 
 type TargetTestType = 'all' | 'passed' | 'failed' | 'skipped';
